- Update to 0.3.0
PR:		ports/78698
Submitted by:	Jeremy Chadwick <freebsd@jdc.parodius.com>

+++ mail/spamass-milter/files/spamass-milter.sh
@@ -24,27 +24,33 @@ rcvar=`set_rcvar`
 command=%%PREFIX%%/sbin/spamass-milter
 required_dirs=%%PREFIX%%/share/spamassassin
 
-# set defaults
-
-spamass_milter_enable=${spamass_milter_enable:-"NO"}
-spamass_milter_socket=${spamass_milter_socket:-"/var/run/spamass-milter.sock"}
-spamass_milter_flags=${spamass_milter_flags:-"-f -p ${spamass_milter_socket}"}
-
 start_postcmd=start_postcmd
 stop_postcmd=stop_postcmd
 
 start_postcmd()
 {
-  sleep 1
-  kill -HUP `head -1 /var/run/sendmail.pid`
+  if test -f /var/run/sendmail.pid
+  then
+	sleep 1
+	kill -HUP `head -1 /var/run/sendmail.pid`
+  fi
 }
 
 stop_postcmd()
 {
   rm -f ${spamass_milter_socket}
-  sleep 1
-  kill -HUP `head -1 /var/run/sendmail.pid`
+  if test -f /var/run/sendmail.pid
+  then
+	sleep 1
+	kill -HUP `head -1 /var/run/sendmail.pid`
+  fi
 }
+
+# set defaults
+
+spamass_milter_enable=${spamass_milter_enable:-"NO"}
+spamass_milter_socket=${spamass_milter_socket:-"/var/run/spamass-milter.sock"}
+spamass_milter_flags=${spamass_milter_flags:-"-f -p ${spamass_milter_socket}"}
 
 load_rc_config ${name}
 run_rc_command "$1"
